backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
dj-database-url==1.3.0
Django==4.2
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-ninja==0.21.0
django-storages==1.13.2
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
gunicorn==20.1.0
idna==3.4
orjson==3.8.10
Pillow==9.5.0
protobuf==4.22.3
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
python-decouple==3.8
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.4.0
urllib3==1.26.15
whitenoise==6.4.0
